refactor: centralize saga utilities
