+++ services/game-session-service/src/main/java/net/firedevops/firemud/service/impl/GameInstanceServiceImpl.java
@@ -5,13 +5,13 @@ import net.firedevops.firemud.client.GameLogicClient;
 import net.firedevops.firemud.common.LoggingUtil;
 import net.firedevops.firemud.common.saga.SagaBuilder;
 import net.firedevops.firemud.common.saga.SagaException;
+import net.firedevops.firemud.common.saga.SagaRunner;
 import net.firedevops.firemud.dto.GameInstanceDto;
 import net.firedevops.firemud.dto.StartSessionRequest;
 import net.firedevops.firemud.entity.GameInstance;
 import net.firedevops.firemud.mapper.GameInstanceMapper;
 import net.firedevops.firemud.repository.GameInstanceRepository;
 import net.firedevops.firemud.service.GameInstanceService;
-import net.firedevops.firemud.service.SagaRunner;
 import net.firedevops.firemud.service.SessionStateService;
 import org.slf4j.Logger;
 import org.springframework.stereotype.Service;
